add python version comparation
